refactor: 탭별 패키지 분리
